+++ docs/changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Synchronize processes that works with file memory and sc-dictionary
 - Get replacements in template result by variable address
 - Configs for file memory string sections
 - Divide strings file into more small files with fixed size by requested configs